[Sri] - Add Nurse name in patient details
